Move Typography to root level
