+++ src/atoms/Heading/Heading.tsx
@@ -2,8 +2,8 @@ import { DetailedHTMLProps, HTMLAttributes } from 'react';
 import { StyledComponentClass } from 'styled-components';
 
 import styled from '_styled-components';
-import { Typography } from 'atoms';
 import Theme, { scale } from 'Theme';
+import Typography from 'Typography';
 
 export const Heading = styled(Typography)`
   color: ${(props: any) => props.as === 'h1' && props.theme.headline};
