+++ src/atoms/Input/Input.tsx
@@ -3,8 +3,8 @@ import { DetailedHTMLProps, InputHTMLAttributes } from 'react';
 import { StyledComponentClass } from 'styled-components';
 
 import styled from '_styled-components';
-import { Typography } from 'atoms';
 import Theme, { borderRadius, scale, transitionDuration } from 'Theme';
+import Typography from 'Typography';
 
 export const Input = styled(Typography)`
   background: ${props => props.theme.controlBackground};
