+++ src/atoms/index.ts
@@ -3,4 +3,3 @@ export { default as Heading } from './Heading';
 export { default as Icon } from './Icon';
 export { default as Input } from './Input';
 export { default as Panel } from './Panel';
-export { default as Typography } from './Typography';
