+++ src/molecules/ActionPanel/ActionPanel.tsx
@@ -7,8 +7,9 @@ import React, {
 import { StyledComponentClass } from 'styled-components';
 
 import styled from '_styled-components';
-import { Panel, Typography } from 'atoms';
+import { Panel } from 'atoms';
 import Theme, { scale } from 'Theme';
+import Typography from 'Typography';
 
 const ActionPanelBody = styled.div`
   ${padding(scale(1), scale(2))};
