+++ src/molecules/List/List.tsx
@@ -2,8 +2,9 @@ import { directionalProperty, padding } from 'polished';
 import React, { Children, ReactNode } from 'react';
 
 import styled from '_styled-components';
-import { Panel, Typography } from 'atoms';
+import { Panel } from 'atoms';
 import { scale } from 'Theme';
+import Typography from 'Typography';
 
 const GroupItem = styled(Typography)`
   ${props =>
